requirements.txt
fastapi
uvicorn[standard]
supabase
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
httpx
pytest
pytest-asyncio
python-dotenv
